Add api auth

+++ app/main.py
@@ -1,16 +1,38 @@
-from flask import Flask, request, jsonify, Response, make_response
-from werkzeug.exceptions import BadRequest
-from sqlalchemy.exc import OperationalError
-from sqlalchemy.exc import DataError
+import datetime
+import json
+import os
 from typing import Tuple
-from app.models import User
-from app.session import get_session
+
+from flask import Flask, request, jsonify, Response, make_response
+import jwt
+from sqlalchemy.exc import OperationalError, DataError, IntegrityError
+from werkzeug.exceptions import BadRequest, HTTPException
+from werkzeug.security import check_password_hash, generate_password_hash
+
+from app.auth import token_required
 from app.functions import replace_skills_with_json, replace_experience_with_json, parse_params, create_param_subs, \
     error_response
+from app.models import User
+from app.session import get_session
 from app.sql_queries import SQL_COUNT, SQL_STATS
 
 app = Flask(__name__)
 app.config['JSON_SORT_KEYS'] = False
+
+
+@app.errorhandler(HTTPException)  # from Flask documentation
+def handle_exception(e):
+    """Return JSON instead of HTML for HTTP errors."""
+    # start with the correct headers and status code from the error
+    response = e.get_response()
+    # replace the body with JSON
+    response.data = json.dumps({
+        "code": e.code,
+        "name": e.name,
+        "description": e.description,
+    })
+    response.content_type = "application/json"
+    return response
 
 
 @app.route('/')
@@ -19,22 +41,56 @@ def index() -> Response:
     return Response(msg, mimetype='text/plain')
 
 
+@app.route('/login')
+def login() -> Tuple[dict, int]:
+    auth = request.authorization
+
+    if not auth or not auth.username or not auth.password:
+        return error_response('not authorized', 401, None)
+
+    session = get_session()
+
+    user = session.query(User).filter_by(username=auth.username).first()
+
+    if not user:
+        return error_response('not authorized', 401, None)
+
+    if check_password_hash(user.password, auth.password):
+        token = jwt.encode(
+            {'username': user.username, 'exp': datetime.datetime.utcnow() + datetime.timedelta(minutes=30)},
+            os.environ['SECRET_KEY'])
+
+        return jsonify({'token': token.decode('UTF-8')})
+
+    return error_response('not authorized', 401, None)
+
+
+@app.route('/identify')
+@token_required
+def identify(current_user: User) -> Tuple[dict, int]:
+    return jsonify(current_user.object_as_dict()), 200
+
+
 @app.route('/api/cv', methods=['GET'])
 @app.route('/api/cv/', methods=['GET'])
-def api_cv_get() -> Response:
+def api_cv_get() -> Tuple[dict, int]:
     # GET ALL CVs
 
     session = get_session()
 
     all_db_records = [user.object_as_dict() for user in session.query(User)]
 
-    return jsonify(all_db_records)
+    return jsonify(all_db_records), 200
 
 
 @app.route('/api/cv', methods=['POST'])
 @app.route('/api/cv/', methods=['POST'])
-def api_cv_post() -> Tuple[dict, int]:
+@token_required
+def api_cv_post(current_user: User) -> Tuple[dict, int]:
     # ADD NEW CV BASED ON JSON DATA
+
+    if not current_user.admin:
+        return error_response('not authorized', 401, None)
 
     session = get_session()
 
@@ -47,8 +103,11 @@ def api_cv_post() -> Tuple[dict, int]:
     new_cv = User()
 
     try:
-        new_cv.username = json_data.get('username', '')
+        new_cv.username = json_data['username']
         new_cv.password = json_data.get('password', '')
+        if new_cv.password != '':
+            new_cv.password = generate_password_hash(new_cv.password, method='sha256', salt_length=8)
+        new_cv.admin = False
         new_cv.firstname = json_data['firstname']
         new_cv.lastname = json_data['lastname']
         replace_skills_with_json(session, new_cv, json_data)
@@ -62,76 +121,116 @@ def api_cv_post() -> Tuple[dict, int]:
 
     try:
         session.commit()
-    except DataError as ex:
+    except (DataError, IntegrityError) as ex:
         return error_response('bad input data', 400, ex)
 
-    return {'success': 'item added'}, 201
+    return jsonify(new_cv.object_as_dict()), 201
 
 
-@app.route('/api/cv/<id>', methods=['GET'])
-def api_cv_id_get(id: int) -> Response:
+@app.route('/api/cv/<user_id>', methods=['GET'])
+def api_cv_id_get(user_id: int) -> Tuple[dict, int]:
     # GET ONE CV OF ID [id]
 
     session = get_session()
 
-    one_db_record = session.query(User).get(id)
+    one_db_record = session.query(User).get(user_id)
 
     try:
-        response = jsonify(one_db_record.object_as_dict())
+        response = jsonify(one_db_record.object_as_dict()), 200
         return response
 
     except AttributeError:
         return make_response('', 404)
 
 
-@app.route('/api/cv/<id>', methods=['PUT'])
-def api_cv_id_put(id: int) -> Tuple[str, int]:
+@app.route('/api/cv/<user_id>', methods=['PUT'])
+@token_required
+def api_cv_id_put(current_user: User, user_id: int) -> Tuple[dict, int]:
     # UPDATE CV OF ID [id] WITH JSON DATA
+
+    if not current_user.admin and current_user.id != int(user_id):
+        return error_response('not authorized', 401, None)
 
     json_data = request.get_json()
 
     session = get_session()
 
-    cv_being_updated = session.query(User).get(id)
+    cv_being_updated = session.query(User).get(user_id)
 
     if cv_being_updated is None:
-        return '', 404
+        return error_response('bad input data', 400, None)
 
-    cv_being_updated.username = json_data.get('username', '')
-    cv_being_updated.password = json_data.get('password', '')
-
+    cv_being_updated.username = json_data['username']
     cv_being_updated.firstname = json_data['firstname']
     cv_being_updated.lastname = json_data['lastname']
 
-    replace_skills_with_json(session, cv_being_updated, json_data)
-
     replace_experience_with_json(cv_being_updated, json_data)
 
-    session.commit()
+    try:
+        replace_skills_with_json(session, cv_being_updated, json_data)
+        session.commit()
+    except IntegrityError as ex:
+        return error_response('bad input data', 400, ex)
 
-    return '', 200
+    return jsonify(cv_being_updated.object_as_dict()), 200
 
 
-@app.route('/api/cv/<id>', methods=['DELETE'])
-def api_cv_id_delete(id: int) -> Tuple[str, int]:
+@app.route('/api/cv/<user_id>', methods=['DELETE'])
+@token_required
+def api_cv_id_delete(current_user: User, user_id: int) -> Tuple[dict, int]:
     # DELETE CV OF ID [id]
+
+    if not current_user.admin:
+        return error_response('not authorized', 401, None)
 
     session = get_session()
 
-    cv_to_be_deleted = session.query(User).get(id)
+    cv_to_be_deleted = session.query(User).get(user_id)
 
     if cv_to_be_deleted is not None:
         session.delete(cv_to_be_deleted)
         session.commit()
 
-        return '', 204
+        return jsonify({}), 204
 
     else:
-        return '', 404
+        return error_response('bad input data', 404, None)
+
+
+@app.route('/api/cv/<user_id>/password', methods=['POST'])
+@token_required
+def api_cv_id_password(current_user: User, user_id: int) -> Tuple[dict, int]:
+    # CHANGE YOUR PASSWORD
+
+    if not current_user.admin and current_user.id != int(user_id):
+        return error_response('not authorized', 401, None)
+
+    json_data = request.get_json()
+
+    try:
+        old_password_from_json = json_data['old_password']
+        new_password_from_json = json_data['new_password']
+    except KeyError as ex:
+        return error_response('bad input data', 400, ex)
+
+    session = get_session()
+
+    cv_being_updated = session.query(User).get(user_id)
+
+    if cv_being_updated is None:
+        return error_response('bad input data', 404, None)
+
+    if check_password_hash(cv_being_updated.password, old_password_from_json):
+        cv_being_updated.password = generate_password_hash(new_password_from_json, method='sha256', salt_length=8)
+        session.commit()
+
+        return jsonify(cv_being_updated.object_as_dict()), 200
+
+    return error_response('bad input data', 400, None)
 
 
 @app.route('/api/cv/stats', methods=['POST'])
-def api_cv_stats() -> Response:
+def api_cv_stats() -> Tuple[dict, int]:
     # GET CVs OF USERS WITH PROVIDED SKILL SET
 
     json_data = request.get_json()
@@ -146,11 +245,11 @@ def api_cv_stats() -> Response:
 
     users_with_skill_set = [{'firstname': user.firstname, 'lastname': user.lastname} for user in result]
 
-    return jsonify(users_with_skill_set)
+    return jsonify(users_with_skill_set), 200
 
 
 @app.route('/api/cv/stats/count', methods=['POST'])
-def api_cv_stats_count() -> Response:
+def api_cv_stats_count() -> Tuple[dict, int]:
     # GET COUNT OF USERS WITH PROVIDED SKILL SET
 
     json_data = request.get_json()
@@ -163,7 +262,7 @@ def api_cv_stats_count() -> Response:
 
     result = session.execute(sql, params).scalar()
 
-    return make_response({'number_of_users_with_skill_set': result})
+    return jsonify({'number_of_users_with_skill_set': result}), 200
 
 
 if __name__ == '__main__':
